Implement save job functionality

+++ Controllers/JobController.js
@@ -93,12 +93,23 @@ const getAllJobs = async (req, res) => {
 const deleteJob = async (req, res) => {
   try {
     const { job_id } = req.params; // Extract job_id from URL params
+    const job = await Job.findById(job_id);
+    const recruiter_id = req.user._id;
+    
+    if(job.recruiter_id !== recruiter_id){
+      return res.status(403).json({ message: "You can only delete job you created!" });
+    }
 
     const job = await Job.findOneAndDelete({ _id: job_id });
 
     if (!job) {
       return res.status(404).json({ message: "Job not found" });
     }
+
+    await User.updateMany(
+      { savedJobs: job_id },
+      { $pull: { savedJobs: job_id } }
+    );
 
     const remainingJobs = await Job.find();
